factor out phone number state from teams add phone screen
